+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,12 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Calculator">
         <activity
+            android:name=".Calc"
+            android:exported="false" />
+        <activity
+            android:name=".media"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
