+++ app/src/main/java/com/example/calculator/media.kt
@@ -1,0 +1,8 @@
+package com.example.calculator
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+class media {
+}
